Cleanup: Moved the BaseManager methods to _utils module
Details:

* Moved the BaseManager methods _divide_filter_args(), _matches_filters((),
  and _matches_prop() as functions into the _utils.py module, in order to
  be able to use them more inependently, and particularly for different
  server-side query properties. This is a preparation for the upcoming
  support for the "List Permitted LPARs / Partitions" operations.
