+++ zhmcclient/_adapter.py
@@ -68,7 +68,8 @@ from ._manager import BaseManager
 from ._resource import BaseResource
 from ._port import PortManager
 from ._logging import logged_api_call
-from ._utils import repr_dict, repr_manager, repr_timestamp
+from ._utils import repr_dict, repr_manager, repr_timestamp, matches_filters, \
+    divide_filter_args
 
 __all__ = ['AdapterManager', 'Adapter']
 
@@ -176,7 +177,8 @@ class AdapterManager(BaseManager):
             resource_obj_list.append(resource_obj)
             # It already has full properties
         else:
-            query_parms, client_filters = self._divide_filter_args(filter_args)
+            query_parms, client_filters = divide_filter_args(
+                self._query_props, filter_args)
 
             resources_name = 'adapters'
             uri = '{}/{}{}'.format(self.cpc.uri, resources_name, query_parms)
@@ -192,7 +194,7 @@ class AdapterManager(BaseManager):
                         name=props.get(self._name_prop, None),
                         properties=props)
 
-                    if self._matches_filters(resource_obj, client_filters):
+                    if matches_filters(resource_obj, client_filters):
                         resource_obj_list.append(resource_obj)
                         if full_properties:
                             resource_obj.pull_full_properties()
