+++ zhmcclient/_capacity_group.py
@@ -30,6 +30,7 @@ import copy
 from ._manager import BaseManager
 from ._resource import BaseResource
 from ._logging import logged_api_call
+from ._utils import matches_filters, divide_filter_args
 
 __all__ = ['CapacityGroupManager', 'CapacityGroup']
 
@@ -117,7 +118,8 @@ class CapacityGroupManager(BaseManager):
             resource_obj_list.append(resource_obj)
             # It already has full properties
         else:
-            query_parms, client_filters = self._divide_filter_args(filter_args)
+            query_parms, client_filters = divide_filter_args(
+                self._query_props, filter_args)
 
             resources_name = 'capacity-groups'
             uri = '{}/{}{}'.format(self.cpc.uri, resources_name, query_parms)
@@ -133,7 +135,7 @@ class CapacityGroupManager(BaseManager):
                         name=props.get(self._name_prop, None),
                         properties=props)
 
-                    if self._matches_filters(resource_obj, client_filters):
+                    if matches_filters(resource_obj, client_filters):
                         resource_obj_list.append(resource_obj)
                         if full_properties:
                             resource_obj.pull_full_properties()
