+++ zhmcclient/_ldap_server_definition.py
@@ -25,6 +25,7 @@ import copy
 from ._manager import BaseManager
 from ._resource import BaseResource
 from ._logging import logged_api_call
+from ._utils import matches_filters, divide_filter_args
 
 __all__ = ['LdapServerDefinitionManager', 'LdapServerDefinition']
 
@@ -116,7 +117,8 @@ class LdapServerDefinitionManager(BaseManager):
           :exc:`~zhmcclient.ConnectionError`
         """
         resource_obj_list = []
-        query_parms, client_filters = self._divide_filter_args(filter_args)
+        query_parms, client_filters = divide_filter_args(
+            self._query_props, filter_args)
         resources_name = 'ldap-server-definitions'
         uri = '{}/{}{}'.format(self.console.uri, resources_name, query_parms)
 
@@ -131,7 +133,7 @@ class LdapServerDefinitionManager(BaseManager):
                     name=props.get(self._name_prop, None),
                     properties=props)
 
-                if self._matches_filters(resource_obj, client_filters):
+                if matches_filters(resource_obj, client_filters):
                     resource_obj_list.append(resource_obj)
                     if full_properties:
                         resource_obj.pull_full_properties()
