+++ zhmcclient/_lpar.py
@@ -34,6 +34,7 @@ from ._manager import BaseManager
 from ._resource import BaseResource
 from ._exceptions import StatusTimeout
 from ._logging import logged_api_call
+from ._utils import matches_filters, divide_filter_args
 
 __all__ = ['LparManager', 'Lpar']
 
@@ -128,7 +129,8 @@ class LparManager(BaseManager):
             resource_obj_list.append(resource_obj)
             # It already has full properties
         else:
-            query_parms, client_filters = self._divide_filter_args(filter_args)
+            query_parms, client_filters = divide_filter_args(
+                self._query_props, filter_args)
 
             resources_name = 'logical-partitions'
             uri = '{}/{}{}'.format(self.cpc.uri, resources_name, query_parms)
@@ -144,7 +146,7 @@ class LparManager(BaseManager):
                         name=props.get(self._name_prop, None),
                         properties=props)
 
-                    if self._matches_filters(resource_obj, client_filters):
+                    if matches_filters(resource_obj, client_filters):
                         resource_obj_list.append(resource_obj)
                         if full_properties:
                             resource_obj.pull_full_properties()
