+++ zhmcclient/_password_rule.py
@@ -26,6 +26,7 @@ import copy
 from ._manager import BaseManager
 from ._resource import BaseResource
 from ._logging import logged_api_call
+from ._utils import matches_filters, divide_filter_args
 
 __all__ = ['PasswordRuleManager', 'PasswordRule']
 
@@ -117,7 +118,8 @@ class PasswordRuleManager(BaseManager):
           :exc:`~zhmcclient.ConnectionError`
         """
         resource_obj_list = []
-        query_parms, client_filters = self._divide_filter_args(filter_args)
+        query_parms, client_filters = divide_filter_args(
+            self._query_props, filter_args)
         resources_name = 'password-rules'
         uri = '{}/{}{}'.format(self.console.uri, resources_name, query_parms)
 
@@ -132,7 +134,7 @@ class PasswordRuleManager(BaseManager):
                     name=props.get(self._name_prop, None),
                     properties=props)
 
-                if self._matches_filters(resource_obj, client_filters):
+                if matches_filters(resource_obj, client_filters):
                     resource_obj_list.append(resource_obj)
                     if full_properties:
                         resource_obj.pull_full_properties()
