+++ zhmcclient/_storage_group_template.py
@@ -33,6 +33,7 @@ from ._manager import BaseManager
 from ._resource import BaseResource
 from ._storage_volume_template import StorageVolumeTemplateManager
 from ._logging import logged_api_call
+from ._utils import matches_filters, divide_filter_args
 
 __all__ = ['StorageGroupTemplateManager', 'StorageGroupTemplate']
 
@@ -136,7 +137,8 @@ class StorageGroupTemplateManager(BaseManager):
             resource_obj_list.append(resource_obj)
             # It already has full properties
         else:
-            query_parms, client_filters = self._divide_filter_args(filter_args)
+            query_parms, client_filters = divide_filter_args(
+                self._query_props, filter_args)
             uri = '{}{}'.format(self._base_uri, query_parms)
 
             result = self.session.get(uri)
@@ -150,7 +152,7 @@ class StorageGroupTemplateManager(BaseManager):
                         name=props.get(self._name_prop, None),
                         properties=props)
 
-                    if self._matches_filters(resource_obj, client_filters):
+                    if matches_filters(resource_obj, client_filters):
                         resource_obj_list.append(resource_obj)
                         if full_properties:
                             resource_obj.pull_full_properties()
