+++ zhmcclient/_storage_volume_template.py
@@ -40,6 +40,7 @@ import copy
 from ._manager import BaseManager
 from ._resource import BaseResource
 from ._logging import logged_api_call
+from ._utils import matches_filters, divide_filter_args
 
 __all__ = ['StorageVolumeTemplateManager', 'StorageVolumeTemplate']
 
@@ -140,7 +141,8 @@ class StorageVolumeTemplateManager(BaseManager):
             resource_obj_list.append(resource_obj)
             # It already has full properties
         else:
-            query_parms, client_filters = self._divide_filter_args(filter_args)
+            query_parms, client_filters = divide_filter_args(
+                self._query_props, filter_args)
 
             resources_name = 'storage-template-volumes'
             uri = '{}/{}{}'.format(self.storage_group_template.uri,
@@ -157,7 +159,7 @@ class StorageVolumeTemplateManager(BaseManager):
                         name=props.get(self._name_prop, None),
                         properties=props)
 
-                    if self._matches_filters(resource_obj, client_filters):
+                    if matches_filters(resource_obj, client_filters):
                         resource_obj_list.append(resource_obj)
                         if full_properties:
                             resource_obj.pull_full_properties()
