+++ zhmcclient/_user_role.py
@@ -31,6 +31,7 @@ import six
 from ._manager import BaseManager
 from ._resource import BaseResource
 from ._logging import logged_api_call
+from ._utils import matches_filters, divide_filter_args
 
 __all__ = ['UserRoleManager', 'UserRole']
 
@@ -121,7 +122,8 @@ class UserRoleManager(BaseManager):
           :exc:`~zhmcclient.ConnectionError`
         """
         resource_obj_list = []
-        query_parms, client_filters = self._divide_filter_args(filter_args)
+        query_parms, client_filters = divide_filter_args(
+            self._query_props, filter_args)
         resources_name = 'user-roles'
         uri = '{}/{}{}'.format(self.console.uri, resources_name, query_parms)
 
@@ -136,7 +138,7 @@ class UserRoleManager(BaseManager):
                     name=props.get(self._name_prop, None),
                     properties=props)
 
-                if self._matches_filters(resource_obj, client_filters):
+                if matches_filters(resource_obj, client_filters):
                     resource_obj_list.append(resource_obj)
                     if full_properties:
                         resource_obj.pull_full_properties()
